requirements.txt
PyQt5
PyAudio
bs4
Pillow
requests

